+++ Thunderstore/CHANGELOG.md
@@ -1,2 +1,15 @@
+### 1.1.0 ###
+- Added UseDashKey config to use a single key for dashing. Rejoining the lobby is required after changing
+- Added DashKeyKeyboard & DashKeyGamepad. Requires a restart after being set
+
+### 1.0.3 ###
+- Fixed bug where settings werent loaded correctly
+
+### 1.0.2 ###
+- Added Enabled / SizeFrom / SizeTo config options
+
+### 1.0.1 ###
+- Updated github repo link
+
 ### 1.0.0 ###
-+ First version
+- First version